requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.18.5
opencv-python>=4.1.1
PyYAML>=5.3.1
scipy>=1.4.1
tqdm>=4.64.0
matplotlib>=3.2.2
tensorboard>=2.4.1
seaborn>=0.11.0
pandas
albumentations>=1.0.3
thop>=0.1.1

